requirements.txt
streamlit
faiss-cpu
numpy
pandas
scikit-learn
python-dotenv
fastapi
uvicorn
torch
tensorflow
tokenizer
sentence-transformers

